refactor: Replace useFreshDate hook with useStoreFreshDate in App component and update related components
